tools: block generator inner transactions (#5506)
